refactor & warning
