+++ easy-to-use-api/klips-url-generator/src/components/ChartComponent/index.tsx
@@ -1,46 +1,54 @@
-import React, { useEffect, useState } from "react";
-import GetCoordinatesString from "./DrawGeometry";
-import { optionsVideoFormat, style } from "../constants";
 import TextArea from "antd/lib/input/TextArea";
+import { optionsBand, style } from "../../constants";
+import GetCoordinatesString from "../DrawGeometry";
+import { CopyOutlined } from '@ant-design/icons';
 import { Button, Input, Tooltip } from "antd";
-import { CopyOutlined } from '@ant-design/icons'
-import { onCopyClickGeom, onCopyClickUrl } from "../service";
+import { onCopyClickGeom, onCopyClickUrl } from "../../service";
+import { useEffect, useState } from "react";
 import OlGeometry from 'ol/geom/Geometry';
+import SelectBand from "../SelectBand";
+import SelectThreshold from "../SelectThreshold";
 import { DrawEvent as OlDrawEvent } from 'ol/interaction/Draw';
-import SelectVideoFormat from "./SelectVideoFormat";
-import WritePersonalTitle from "./WritePersonalTitle";
 
-export interface VideoProps {
+
+export interface ChartProps {
     geoJsonGeom: string,
     region: string;
+    wktGeom: string;
     onDrawEnd?: (geom: OlGeometry) => void;
     onDrawStart?: (event: OlDrawEvent) => void;
 };
 
-export type VideoComponentProps = VideoProps;
+export type ChartComponentProps = ChartProps;
 
-const VideoComponent: React.FC<VideoComponentProps> = ({onDrawEnd, onDrawStart, geoJsonGeom, region}) => {
+const ChartComponent: React.FC<ChartComponentProps> = ({ geoJsonGeom, region, wktGeom, onDrawEnd, onDrawStart }) => {
+    const [band, setBand] = useState('');
+    const [threshold, setThreshold] = useState();
     const [url, setURL] = useState('');
-    const [videoFormat, setVideoFormat] = useState('');
-    const [personalTitle, setPersonalTitle] = useState('');
 
     useEffect(() => {
-        if (region && geoJsonGeom && videoFormat && personalTitle) {
-            setURL(`https://klips-dev.terrestris.de/?region=${region.toLowerCase()}&area-of-interest=${geoJsonGeom}&output=${videoFormat}&title=${personalTitle}`)
+        if (region && wktGeom && threshold && band) {
+            setURL(`https://klips-dev.terrestris.de/?region=${region.toLowerCase()}&geom=${wktGeom}&threshold=${threshold}&band=${band}`)
         };
-    }, [region, geoJsonGeom, videoFormat, personalTitle]);
+    }, [region, wktGeom, threshold, band]);
 
-    const changeVideoFormat = (newFormat: string) => {
-        setVideoFormat(newFormat);
+    const changeBand = (newBand: string) => {
+        setBand(newBand);
     };
 
-    const changeTitle = (newTitle: string) => {
-        setPersonalTitle(newTitle);
+    const changeThreshold = (newThreshold: any) => {
+        setThreshold(newThreshold);
     };
 
     return (
-        <div className='video-component'>
+        <>
             <div className='geometry'>
+                <GetCoordinatesString
+                    drawType='Point'
+                    drawStyle={style.point}
+                    onDrawEnd={onDrawEnd}
+                    onDrawStart={onDrawStart}
+                />
                 <GetCoordinatesString
                     drawType='Polygon'
                     drawStyle={style.polygon}
@@ -66,13 +74,13 @@ const VideoComponent: React.FC<VideoComponentProps> = ({onDrawEnd, onDrawStart, 
                 }
             </div>
             <div className='attributes'>
-                <SelectVideoFormat
-                    selectedVideoFormat={videoFormat}
-                    changeVideoFormat={changeVideoFormat}
-                    inputVideoFormats={optionsVideoFormat}
+                <SelectBand
+                    inputBands={optionsBand}
+                    changeBand={changeBand}
+                    selectedBand={band}
                 />
-                <WritePersonalTitle
-                    changeTitle={changeTitle}
+                <SelectThreshold
+                    changeThreshold={changeThreshold}
                 />
             </div>
             <div className='permalink'>
@@ -90,8 +98,16 @@ const VideoComponent: React.FC<VideoComponentProps> = ({onDrawEnd, onDrawStart, 
                     />
                 </Tooltip>
             </div>
-        </div >
-    );
+        </>
+    )
 };
 
-export default VideoComponent;
+export default ChartComponent;
+
+
+
+
+
+
+
+
